requirements.txt
# MRKL Agent Dependencies

# Core LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# API Framework
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
